Added SHA-1 algorithm

+++ src/main/java/men/brakh/cryptyhash/CryptoHash.java
@@ -1,6 +1,8 @@
 package men.brakh.cryptyhash;
 
+import java.math.BigInteger;
+
 public interface CryptoHash {
     String getHash(byte[] msg);
-    int getIntHash(byte[] msg);
+    BigInteger getIntHash(byte[] msg);
 }
